Move failure exit from reporter -> runner
In case of failures there was a leftover 'process.exit(1)' in a reporter
implementation that shouldn't be there. With this it is now moved to
Runner which will exit if at least 1 test has failed *and* there is no
one listening on 'finish' event (the one who is listening is responsible
for exiting then).
'finish' event now contains hasFailures (boolean) argument.

PR-URL: https://github.com/metarhia/metatests/pull/79
Reviewed-By: Timur Shemsedinov <timur.shemsedinov@gmail.com>
Reviewed-By: Dmytro Nechai <nechaido@gmail.com>