fix: remove yaml warnings
Due to the fact that newer versions of yaml library (1.7.0+)
produce warning when trying to print nested objects in `set`
and at the same time tap-mocha-reporter doesn't support yaml
output that fixes this warning (mapAsMap: true) we have to just
silence them like this.
This would still result in "empty" diff on some reporters but
at least we will not get warnings that clutter whole report.

Refs: https://github.com/metarhia/metatests/issues/265